Implement autofix for E713 and E714
