+++ README.md
@@ -359,10 +359,10 @@ For more, see [pycodestyle](https://pypi.org/project/pycodestyle/2.9.1/) on PyPI
 | ---- | ---- | ------- | --- |
 | E402 | ModuleImportNotAtTopOfFile | Module level import not at top of file |  |
 | E501 | LineTooLong | Line too long (89 > 88 characters) |  |
-| E711 | NoneComparison | Comparison to `None` should be `cond is None` |  |
-| E712 | TrueFalseComparison | Comparison to `True` should be `cond is True` |  |
-| E713 | NotInTest | Test for membership should be `not in` |  |
-| E714 | NotIsTest | Test for object identity should be `is not` |  |
+| E711 | NoneComparison | Comparison to `None` should be `cond is None` | 🛠 |
+| E712 | TrueFalseComparison | Comparison to `True` should be `cond is True` | 🛠 |
+| E713 | NotInTest | Test for membership should be `not in` | 🛠 |
+| E714 | NotIsTest | Test for object identity should be `is not` | 🛠 |
 | E721 | TypeComparison | Do not compare types, use `isinstance()` |  |
 | E722 | DoNotUseBareExcept | Do not use bare `except` |  |
 | E731 | DoNotAssignLambda | Do not assign a lambda expression, use a def |  |
